Relicensed under GPL v3.0 or later
Previously, code in `src/` and `search/` was licensed under Apache 2.0.
The reason for the new license is that the search component uses Xapian
library that is distributed under GPL v2.0 or later. So licensing search
under Apache 2.0 was a mistake since Apache 2.0 may be used in a GPL v3.0
project, and not vice versa. The code in `src/` does not have this
limitation, but it is relicensed under GPL v3.0 or later for consistency
with the search component.