Modified directory layout
Renamed back the subrepository "content" to "docs". Created directory
"app/content/" with symlinks to "content". The new layout is better since
the app and search backend will not scan the entire subrepository.
In the old layout, the app and search may touch files that should never
be served or searched.